moving socket.io files inside client
